2.ChatModels/3_chatmodel_google.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv
load_dotenv()

model = ChatGoogleGenerativeAI(model='gemini-1.5-pro')
result = model.invoke("which last date you should have asnwer means what was the latest content you hahve delevered ");
print(result.content)